requirements.txt
pandas==2.0.3
transformers==4.38.2
omegaconf==2.3.0
pandas==2.0.3
scikit-learn==1.3.0
tqdm==4.65.0
dataframe_image==0.2.1
spacy==2.3.9
nltk==3.6.7
hydra-core==1.3.0
tensorboard==2.14.0
jupyterlab
